+++ activedb-core/src/engine/tests/vector_tests.rs
@@ -3,7 +3,7 @@ use crate::engine::vector_core::vector_distance::{MAX_DISTANCE, MIN_DISTANCE, OR
 use crate::engine::vector_core::vector::HVector;
 use bumpalo::Bump;
 
-fn alloc_vector<'a>(arena: &'a Bump, data: &[f64]) -> HVector<'a> {
+fn alloc_vector<'a>(arena: &'a Bump, data: &[f32]) -> HVector<'a> {
     let slice = arena.alloc_slice_copy(data);
     HVector::from_slice("vector", 0, slice)
 }
